fix(#105): compare cash reversal amounts as bigint in pgTAP
pos_signed_minor is a bigint domain, so the exact-reversal assertion has to cast before pgTAP is() can compare it.

+++ supabase/tests/admin_cash_correction.sql
@@ -58,8 +58,8 @@ SELECT lives_ok(
 );
 
 SELECT is(
-  (SELECT signed_amount_minor FROM pos_cash_movements WHERE corrects_movement_id = '88888888-8888-4888-8888-888888888881'),
-  -2500,
+  (SELECT signed_amount_minor::bigint FROM pos_cash_movements WHERE corrects_movement_id = '88888888-8888-4888-8888-888888888881'),
+  -2500::bigint,
   'correction amount is the exact reverse'
 );
